requirements.txt
requests==2.21.0
python-dateutil==2.6.1
